backend/requirements.txt
# FastAPI and Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.31
psycopg==3.2.9
psycopg-binary==3.2.9

# Pydantic for data validation and settings
pydantic[email]==2.8.2
pydantic-settings==2.3.4

# Security and Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv==1.0.1